implement MainViewModel
