+++ app/src/main/java/com/submission/appstory/viewModel/MainViewModel.kt
@@ -1,0 +1,50 @@
+package com.submission.appstory.viewModel
+
+import android.util.Log
+import android.view.View
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import com.submission.appstory.MainActivity
+import com.submission.appstory.api.ApiConfig
+import com.submission.appstory.response.ListStoryItem
+import com.submission.appstory.response.StoriesResponse
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
+
+class MainViewModel: ViewModel() {
+    private val _isLoading = MutableLiveData<Boolean>()
+    val isLoading: LiveData<Boolean> = _isLoading
+
+    private val _listStory = MutableLiveData<List<ListStoryItem>>()
+    val listStory: LiveData<List<ListStoryItem>> = _listStory
+
+    fun getStories(token: String) {
+        _isLoading.value = true
+        val call = ApiConfig.getApiService(token).getAllStories()
+        call.enqueue(object: Callback<StoriesResponse> {
+            override fun onResponse(
+                call: Call<StoriesResponse>,
+                response: Response<StoriesResponse>
+            ) {
+                if (response.isSuccessful) {
+                    _isLoading.value = false
+                    val responseBody = response.body()!!
+                    _listStory.value = responseBody.listStory
+                } else {
+                    Log.e(TAG, "onFailure: ${response.message()}")
+                }
+            }
+
+            override fun onFailure(call: Call<StoriesResponse>, t: Throwable) {
+                _isLoading.value = false
+                Log.e(TAG, "onFailure: ${t.message}")
+            }
+        })
+    }
+
+    companion object {
+        private const val TAG = "MainActivity"
+    }
+}
